feat(cashier): add execute transaction shortcut and enhance payment input
- Introduced 'execute-transaction' keyboard shortcut to process payments directly.
- Added 'Execute Transaction' entry in settings localization.
- Created PaymentInput component to manage payment amount input and interactions.
- Refactored PaymentDialog to utilize PaymentInput for better separation of concerns.
- Removed shortcut feedback hooks as they were no longer needed.
